add user id verification from request

+++ src/controllers/table.controller.js
@@ -4,9 +4,8 @@ const { response } = require('../utils/response');
 
 exports.get = async (req, res, next) => {
 	try {
-		res.json(await tableService.get());
+		return res.json(await tableService.get(req.user.id));
 	} catch (err) {
-		console.error(err.message);
 		next(err);
 	}
 };
@@ -14,38 +13,43 @@ exports.create = async (req, res, next) => {
 	req.body.user_id = req.user.id;
 	try {
 		const validated = validationResult(req);
-		if (!validated.isEmpty()) res.json(response(400, 'Harap masukan data dengan benar', validated));
+		if (!validated.isEmpty()) return res.json(response(400, 'Harap masukan data dengan benar', validated));
 
-		res.json(await tableService.create(req.body));
+		return res.json(await tableService.create(req.body));
 	} catch (err) {
-		console.error(err.message);
 		next(err);
 	}
 };
 exports.show = async (req, res, next) => {
 	try {
-		res.json(await tableService.show(parseInt(req.params.id)));
+		const validateByUserId = await tableService.getByUserId(parseInt(req.params.id), req.user.id);
+		if (validateByUserId.payload.status != 200) return res.json(validateByUserId);
+
+		return res.json(await tableService.show(parseInt(req.params.id)));
 	} catch (err) {
-		console.error(err.message);
 		next(err);
 	}
 };
 exports.update = async (req, res, next) => {
 	try {
 		const validated = validationResult(req);
-		if (!validated.isEmpty()) res.json(response(400, 'Harap masukan data dengan benar', validated));
+		if (!validated.isEmpty()) return res.json(response(400, 'Harap masukan data dengan benar', validated));
 
-		res.json(await tableService.update(parseInt(req.params.id), req.body));
+		const validateByUserId = await tableService.getByUserId(parseInt(req.params.id), req.user.id);
+		if (validateByUserId.payload.status != 200) return res.json(validateByUserId);
+
+		return res.json(await tableService.update(parseInt(req.params.id), req.body));
 	} catch (err) {
-		console.error(err.message);
 		next(err);
 	}
 };
 exports.destroy = async (req, res, next) => {
 	try {
-		res.json(await tableService.destroy(parseInt(req.params.id)));
+		const validateByUserId = await tableService.getByUserId(parseInt(req.params.id), req.user.id);
+		if (validateByUserId.payload.status != 200) return res.json(validateByUserId);
+
+		return res.json(await tableService.destroy(parseInt(req.params.id)));
 	} catch (err) {
-		console.error(err.message);
 		next(err);
 	}
 };
@@ -53,7 +57,10 @@ exports.destroy = async (req, res, next) => {
 exports.validate = (method) => {
 	switch (method) {
 		case 'createOrUpdate': {
-			return [check('table_data', 'bidang table_data harus di isi').exists(), check('table_number', 'bidang table_number harus di isi').exists()];
+			return [
+				check('table_id', 'bidang table_id harus di isi').exists(),
+				check('table_data', 'bidang table_data harus di isi').exists(),
+			];
 		}
 	}
 };
